==> 1_preparation/01_report_data_organization.Rmd <==
---
title: "Report_data_organization"
author: "Yuki Saito"
date: "2024-03-05"
output: html_document
---

- R markdown script to process and organize report data downloaded from C-CAT. 

```{r setup, include=FALSE, message=FALSE, warning=FALSE}
knitr::opts_chunk$set(echo = TRUE)
knitr::opts_knit$set(root.dir = rprojroot::find_rstudio_root_file())
knitr::opts_chunk$set(root.dir = rprojroot::find_rstudio_root_file())
```

```{r library, message=FALSE, warning=FALSE}
library(tidyverse)
library(dplyr)
library(readxl)
```

```{r make directory}
dir.create("result/table/prep/report", showWarnings = FALSE, recursive = TRUE)
```

```{r preparation, warning=FALSE, message=FALSE}
organ <- "bowel"
files <- list.files(paste0("../../../sh_c-cat_analysis/data/ccat/ver20240621/F1CDx/", organ, "/"), 
                    pattern = "^report.*\\.csv$", 
                    full.names = TRUE, 
                    recursive = TRUE)
report_data <- lapply(files, 
                      function(file) 
                        read_csv(file,
                                 locale = locale(encoding = "Shift-JIS")))

col_types <- sapply(report_data[[11]], class)

col_types_tibble <- tibble(
  Column = names(col_types),
  Type = as.character(col_types)
)

col_types_tibble %>% 
  write_excel_csv("result/table/prep/report/col_types.csv", na = "")
```

```{r preparation}
# list up organ names
organ_names <- list.files("../../../sh_c-cat_analysis/data/ccat/ver20240621/F1CDx/",
                          full.names = FALSE)
organ_names[1:3]

# colnames reference
report_colnames_df <- read_xlsx("../../ref/colnames_ver202406.xlsx", "report")
head(report_colnames_df)

report_cols <- report_colnames_df
report_cols["coltypes"][report_cols["coltypes"] == "character"] <- "c"
report_cols["coltypes"][report_cols["coltypes"] == "numeric"] <- "n"
report_cols["coltypes"][report_cols["coltypes"] == "logical"] <- "l"
report_cols["coltypes"][report_cols["coltypes"] == "date"] <- "D"
report_cols_collapsed <- paste(report_cols$coltypes, collapse = "")

# column separation
number_columns <- report_colnames_df %>% 
  filter(Number_columns) %>%
  select(colnames_ys)
alteration_columns <- report_colnames_df %>% 
  filter(Alteration_columns) %>%
  select(colnames_ys)
actionability_columns <- report_colnames_df %>% 
  filter(Actionability_columns) %>%
  select(colnames_ys)
```

```{r analysis, warning=FALSE, message=FALSE}
for (organ in organ_names){
  # list all csv files in the directory, including subdirectories
  files <- list.files(paste0("../../../sh_c-cat_analysis/data/ccat/ver20240621/F1CDx/", organ, "/"), 
                      pattern = "^report.*\\.csv$", 
                      full.names = TRUE, 
                      recursive = TRUE)
  # read all these csv files (UTF-8 encoding is used by default)) 
  report_data <- lapply(files, 
                        function(file) 
                          read_csv(file, 
                                   locale = locale(encoding = "Shift-JIS"), 
                                   col_types = report_cols_collapsed))
  
  # separate data
  number_data <- list()
  alteration_data <- list()
  actionability_data <- list()
  
  for (i in 1:length(report_data)){
    # check if column names match
    if (all(colnames(report_data[[i]]) != report_colnames_df$names_EN)) {
      stop("Column names do not match")
    }
    # rename columns
    colnames(report_data[[i]]) <- report_colnames_df$colnames_ys
    
    # select columns
    number_data[[i]] <- report_data[[i]] %>% 
      select(number_columns$colnames_ys) %>% 
      unique()
    alteration_data[[i]] <- report_data[[i]] %>%
      select(alteration_columns$colnames_ys) %>% 
      unique()
    actionability_data[[i]] <- report_data[[i]] %>%
      select(actionability_columns$colnames_ys) %>% 
      unique()
  }
  
  # merge all data
  number_data <- bind_rows(number_data)
  alteration_data <- bind_rows(alteration_data)
  actionability_data <- bind_rows(actionability_data)
  rm(report_data)

  # make directiory if not exist
  dir.create(paste0("result/table/prep/report/", organ), showWarnings = FALSE, recursive = TRUE)
  
  # check
  tmp <- alteration_data %>% 
    filter(!Panel_name %in% c("FoundationOne CDx DX1", "FoundationOne CDx DX2"))
  if (nrow(tmp) > 0){
    tmp$Panel_name %>% 
      unique() %>% 
      print()
    tmp %>% 
      select(Tumor_Sample_Barcode, Panel_name) %>% 
      write_excel_csv(paste0("result/table/prep/report/", organ, "/non_FoundationOne_data.csv"), na = "")
  }
  
  # panel selection
  number_data <- number_data %>% 
    filter(Panel_name %in% c("FoundationOne CDx DX1", "FoundationOne CDx DX2"))
  alteration_data <- alteration_data %>% 
    filter(Panel_name %in% c("FoundationOne CDx DX1", "FoundationOne CDx DX2"))
  actionability_data <- actionability_data %>% 
    filter(Panel_name %in% c("FoundationOne CDx DX1", "FoundationOne CDx DX2"))
  
  # save as tsv
  # alteration_data
  alteration_data %>%
    filter(Alteration_type == "copy_number_alteration") %>% 
    select(-c("TMB", "TMB_total", "MSI", "LOH")) %>% 
    write_excel_csv(paste0("result/table/prep/report/", organ, "/CNV_data.csv"), na = "")
  alteration_data %>% 
    filter(Alteration_type == "interaction") %>% 
    write_excel_csv(paste0("result/table/prep/report/", organ, "/interaction_data.csv"), na = "")
  alteration_data %>% 
    filter(Alteration_type == "rearrangement") %>% 
    select(-c("TMB", "TMB_total", "MSI", "LOH")) %>%
    write_excel_csv(paste0("result/table/prep/report/", organ, "/rearrangement_data.csv"), na = "")
  alteration_data %>% 
    filter(Alteration_type == "small_scale_variant") %>% 
    select(-c("TMB", "TMB_total", "MSI", "LOH")) %>% 
    write_excel_csv(paste0("result/table/prep/report/", organ, "/small_scale_variant_data.csv"), na = "")
  alteration_data %>% 
    filter(Alteration_type == "other_biomarker") %>% 
    select("Tumor_Sample_Barcode", "Cancer_type", "Cancer_type_last_level", "Specimen_collection_date", "Panel_name",
           "Alteration_type", "Somatic_or_Germline", "DNA_or_RNA", "Marker", "Mutation_context", "Marker_detail",
           "TMB", "TMB_total", "MSI", "LOH", "Database_update_date") %>%
    write_excel_csv(paste0("result/table/prep/report/", organ, "/other_biomarker_data.csv"), na = "")
  
  # check
  other <- alteration_data %>% 
    filter(!Alteration_type %in% c("copy_number_alteration", "interaction", "rearrangement", "small_scale_variant", "other_biomarker")) %>% 
    nrow()
  if (other > 0){
   stop("There are other types of alteration.")
  }
  
  number_data %>% 
    write_excel_csv(paste0("result/table/prep/report/", organ, "/number_data.csv"), na = "")
  actionability_data %>%
    write_excel_csv(paste0("result/table/prep/report/", organ, "/actionability_data.csv"), na = "")
}
```

```{r sessioninfo}
sessionInfo()
```
