==> 1_preparation/03_patient_data.Rmd <==
---
title: "Patient_data"
author: "Yuki Saito"
date: "2024-03-05"
output: html_document
---

- R markdown script to combine patient data into one file. 

```{r setup, include=FALSE, message=FALSE, warning=FALSE}
knitr::opts_chunk$set(echo = TRUE)
knitr::opts_knit$set(root.dir = rprojroot::find_rstudio_root_file())
knitr::opts_chunk$set(root.dir = rprojroot::find_rstudio_root_file())
```

```{r library, message=FALSE, warning=FALSE}
library(tidyverse)
library(dplyr)
library(readxl)
```

```{r make directory}
dir.create("result/table/merged", showWarnings = FALSE)
```

```{r preparation}
# list up organ names
organ_names <- list.files("../../../sh_c-cat_analysis/data/ccat/ver20240621/F1CDx/",
                          full.names = FALSE)
organ_names[1:3]

# colnames reference
case_colnames_df <- read_xlsx("../../ref/colnames_ver202406.xlsx", "case")
head(case_colnames_df)

patient_column_df <- case_colnames_df %>% 
  filter(Patient_columns == "TRUE")

report_cols <- patient_column_df
report_cols["coltypes"][report_cols["coltypes"] == "character"] <- "c"
report_cols["coltypes"][report_cols["coltypes"] == "numeric"] <- "n"
report_cols["coltypes"][report_cols["coltypes"] == "logical"] <- "l"
report_cols["coltypes"][report_cols["coltypes"] == "date"] <- "D"
report_cols_collapsed <- paste(report_cols$coltypes, collapse = "")
```

```{r analysis, warning=FALSE, message=FALSE}
# read in case files
patient_data <- list()
for (organ in organ_names){
  organ_patient_data <- read_csv(paste0("result/table/prep/case/", organ, "/patient_data.csv"),
                                 col_types = report_cols_collapsed)
  patient_data[[organ]] <- organ_patient_data
}
patient_data <- bind_rows(patient_data)

# if there are duplicated patient data, stop
duplicated_patient_data <- patient_data %>%
  count(Tumor_Sample_Barcode) %>%
  filter(n >= 2)
if (nrow(duplicated_patient_data) > 0){
  print(duplicated_patient_data)
  stop("There are duplicated patient data.")
}

# check the number of patient data
print("Number of patients")
nrow(patient_data)

# save the merged patient data
patient_data %>% 
  write_excel_csv("result/table/merged/merged_patient_data.csv", na = "")
```

```{r sessioninfo}
sessionInfo()
```
