==> 1_preparation/05_actionability_data.Rmd <==
---
title: "Actionability_data"
author: "Yuki Saito"
date: "2024-03-06"
output: html_document
---

- R markdown script to annotate clinical actionability.

```{r setup, include=FALSE, message=FALSE, warning=FALSE}
knitr::opts_chunk$set(echo = TRUE)
knitr::opts_knit$set(root.dir = rprojroot::find_rstudio_root_file())
knitr::opts_chunk$set(root.dir = rprojroot::find_rstudio_root_file())
```

```{r library, message=FALSE, warning=FALSE}
library(tidyverse)
library(dplyr)
library(readxl)
```

```{r source}
source("../../R/utils.R")
source("../../R/figure.R")
```

```{r make directory}
dir.create("result/table/merged", showWarnings = FALSE)
```

```{r function}
collapse_without_na <- function(x) {
  paste(na.omit(x), collapse = ":")
}
```

```{r preparation}
# list up organ names
organ_names <- list.files("../../../sh_c-cat_analysis/data/ccat/ver20240621/F1CDx/",
                          full.names = FALSE)
organ_names[1:3]

# colnames reference
report_colnames_df <- read_xlsx("../../ref/colnames_ver202406.xlsx", "report") %>% 
  filter(Actionability_columns == "TRUE")
head(report_colnames_df)

report_cols <- report_colnames_df
report_cols["coltypes"][report_cols["coltypes"] == "character"] <- "c"
report_cols["coltypes"][report_cols["coltypes"] == "numeric"] <- "n"
report_cols["coltypes"][report_cols["coltypes"] == "logical"] <- "l"
report_cols["coltypes"][report_cols["coltypes"] == "date"] <- "D"
report_cols_collapsed <- paste(report_cols$coltypes, collapse = "")
```

```{r analysis, warning=FALSE, message=FALSE}
actionability_data <- list()
for (organ in organ_names){
  # read in actionability files
  organ_actionability_data <- read_csv(paste0("result/table/prep/report/", organ, "/actionability_data.csv"),
                                       col_types = report_cols_collapsed)
  
  # filter out NA evidence level & column selection
  organ_actionability_data <- organ_actionability_data %>% 
    filter(!is.na(Evidence_level)) %>% 
    mutate(Alteration = paste(Marker, Mutation_context, Somatic_or_Germline, sep = "_")) %>%
    select(Tumor_Sample_Barcode,
           Cancer_type,
           Cancer_type_last_level,
           Specimen_collection_date,
           Panel_name,
           Alteration,
           Marker_detail,
           Evidence_level,
           DNA_or_RNA,
           Marker_detail,
           Drug,
           Reachability_to_drugs,
           Database_update_date)
  
  actionability_data[[organ]] <- organ_actionability_data
}
actionability_data <- bind_rows(actionability_data)
```

```{r actionability check}
actionability_data %>%
  count(Evidence_level)

actionability_data %>% 
  filter(Evidence_level == "F") %>% 
  count(Drug)

actionability_data %>% 
  filter(Evidence_level == "F") %>% 
  filter(!is.na(Drug))

actionability_data %>% 
  filter(Tumor_Sample_Barcode == "C18A20F4D961465BB0A7BF09C95B98621EE75E2C602C7EF3F237CD0566BEFBDE") %>% 
  filter(Alteration == "PTEN_loss_Somatic") %>% 
  select(Evidence_level, Drug)

actionability_data %>% 
  filter(Evidence_level != "F") %>% 
  filter(is.na(Drug))

actionability_data %>% 
  filter(str_detect(Drug, ":") | str_detect(Drug, ";"))

# resistance
actionability_data %>% 
  filter(Evidence_level == "R1") %>% 
  count(Alteration, Cancer_type, sort = TRUE)
actionability_data %>% 
  filter(Evidence_level == "R1*") %>%
  count(Alteration, Cancer_type, sort = TRUE)
actionability_data %>% 
  filter(Evidence_level == "R2") %>% 
  count(Alteration, Cancer_type, sort = TRUE)
actionability_data %>% 
  filter(Evidence_level == "R2*") %>% 
  count(Alteration, Cancer_type, sort = TRUE)
actionability_data %>% 
  filter(Evidence_level == "R3") %>% 
  count(Alteration, Cancer_type, sort = TRUE)
actionability_data %>% 
  filter(Evidence_level == "R3*") %>% 
  count(Alteration, Cancer_type, sort = TRUE)
```

```{r analysis2, warning=FALSE, message=FALSE}
actionability_data <- list()
actionability_data2 <- list()

for (organ in organ_names){
  # read in actionability files
  organ_actionability_data <- read_csv(paste0("result/table/prep/report/", organ, "/actionability_data.csv"),
                                       col_types = report_cols_collapsed)
  
  # select evidence levels A-E
  organ_actionability_data <- organ_actionability_data %>% 
    as_tibble() %>% 
    filter(!is.na(Evidence_level)) %>% 
    filter(Evidence_level %in% c("A", "B", "C", "D", "E")) %>% 
    mutate(Alteration = paste(Marker, Mutation_context, Somatic_or_Germline, sep = "_")) %>%
    mutate(Drug_reachability = paste0(Drug, "(", Reachability_to_drugs, ")")) %>% 
    select(Tumor_Sample_Barcode,
           Cancer_type,
           Cancer_type_last_level,
           Converted_cancer_type,
           Converted_cancer_type_selected,
           Specimen_collection_date,
           Panel_name,
           DNA_or_RNA,
           Alteration,
           Marker_detail,
           Evidence_level,
           Drug,
           Drug_reachability,
           Database_update_date) %>% 
    unique()
  
  # pivot wider for evidence levels
  organ_actionability_wide_data <- organ_actionability_data %>% 
    select(-Drug_reachability) %>%
    pivot_wider(names_from = Evidence_level,
                values_from = Drug)
  organ_actionability_wide_reachability_data <- organ_actionability_data %>%
    mutate(Evidence_level2 = paste0(Evidence_level, "_reachability")) %>% 
    select(Tumor_Sample_Barcode, Alteration, Evidence_level2, Drug_reachability) %>%
    pivot_wider(names_from = Evidence_level2,
                values_from = Drug_reachability)
  
  new_cols <- c("A", "B", "C", "D", "E")
  for (col in new_cols) {
    if (!(col %in% colnames(organ_actionability_wide_data))) {
      organ_actionability_wide_data <- organ_actionability_wide_data %>% 
        mutate(!!col := NA_character_)
    }
  }
  
  new_cols2 <- c("A_reachability", "B_reachability", "C_reachability", "D_reachability", "E_reachability")
  for (col in new_cols2){
    if (!(col %in% colnames(organ_actionability_wide_reachability_data))) {
      organ_actionability_wide_reachability_data <- organ_actionability_wide_reachability_data %>% 
        mutate(!!col := NA_character_)
    }
  }
  
  organ_actionability_wide_data2 <- organ_actionability_wide_data %>% 
    mutate(Evidence_A = sapply(A, collapse_without_na),
           Evidence_B = sapply(B, collapse_without_na),
           Evidence_C = sapply(C, collapse_without_na),
           Evidence_D = sapply(D, collapse_without_na),
           Evidence_E = sapply(E, collapse_without_na)) %>%
    select("Tumor_Sample_Barcode",
           "Cancer_type",
           "Cancer_type_last_level",
           "Specimen_collection_date",
           "Panel_name",
           "Database_update_date",
           "Alteration",
           "Evidence_A",
           "Evidence_B",
           "Evidence_C",
           "Evidence_D",
           "Evidence_E")  
  organ_actionability_wide_reachability_data2 <- organ_actionability_wide_reachability_data %>% 
    mutate(Evidence_A_reachability = sapply(A_reachability, collapse_without_na),
           Evidence_B_reachability = sapply(B_reachability, collapse_without_na),
           Evidence_C_reachability = sapply(C_reachability, collapse_without_na),
           Evidence_D_reachability = sapply(D_reachability, collapse_without_na),
           Evidence_E_reachability = sapply(E_reachability, collapse_without_na)) %>%
    select("Tumor_Sample_Barcode",
           "Alteration",
           "Evidence_A_reachability",
           "Evidence_B_reachability",
           "Evidence_C_reachability",
           "Evidence_D_reachability",
           "Evidence_E_reachability")
  organ_actionability_wide_data2 <- organ_actionability_wide_data2 %>% 
    left_join2(organ_actionability_wide_reachability_data2)
  
  organ_actionability_wide_data3 <- organ_actionability_wide_data %>% 
    mutate(Evidence_A = sapply(A, collapse_without_na),
           Evidence_B = sapply(B, collapse_without_na),
           Evidence_C = sapply(C, collapse_without_na),
           Evidence_D = sapply(D, collapse_without_na),
           Evidence_E = sapply(E, collapse_without_na)) %>%
    mutate(Evidence_A = if_else(Evidence_A != "", paste(Alteration, Evidence_A, sep = ";"), NA_character_),
           Evidence_B = if_else(Evidence_B != "", paste(Alteration, Evidence_B, sep = ";"), NA_character_),
           Evidence_C = if_else(Evidence_C != "", paste(Alteration, Evidence_C, sep = ";"), NA_character_),
           Evidence_D = if_else(Evidence_D != "", paste(Alteration, Evidence_D, sep = ";"), NA_character_),
           Evidence_E = if_else(Evidence_E != "", paste(Alteration, Evidence_E, sep = ";"), NA_character_)) %>% 
    select("Tumor_Sample_Barcode",
           "Cancer_type",
           "Cancer_type_last_level",
           "Specimen_collection_date",
           "Panel_name",
           "Database_update_date",
           "Evidence_A",
           "Evidence_B",
           "Evidence_C",
           "Evidence_D",
           "Evidence_E")
  organ_actionability_wide_reachability_data3 <- organ_actionability_wide_reachability_data %>% 
    mutate(Evidence_A_reachability = sapply(A_reachability, collapse_without_na),
           Evidence_B_reachability = sapply(B_reachability, collapse_without_na),
           Evidence_C_reachability = sapply(C_reachability, collapse_without_na),
           Evidence_D_reachability = sapply(D_reachability, collapse_without_na),
           Evidence_E_reachability = sapply(E_reachability, collapse_without_na)) %>%
    mutate(Evidence_A_reachability = if_else(Evidence_A_reachability != "", paste(Alteration, Evidence_A_reachability, sep = ";"), NA_character_),
           Evidence_B_reachability = if_else(Evidence_B_reachability != "", paste(Alteration, Evidence_B_reachability, sep = ";"), NA_character_),
           Evidence_C_reachability = if_else(Evidence_C_reachability != "", paste(Alteration, Evidence_C_reachability, sep = ";"), NA_character_),
           Evidence_D_reachability = if_else(Evidence_D_reachability != "", paste(Alteration, Evidence_D_reachability, sep = ";"), NA_character_),
           Evidence_E_reachability = if_else(Evidence_E_reachability != "", paste(Alteration, Evidence_E_reachability, sep = ";"), NA_character_)) %>% 
    select("Tumor_Sample_Barcode",
           "Evidence_A_reachability",
           "Evidence_B_reachability",
           "Evidence_C_reachability",
           "Evidence_D_reachability",
           "Evidence_E_reachability")
  
  organ_actionability_wide_data3 <- organ_actionability_wide_data3 %>%
    group_by(Tumor_Sample_Barcode,
             Cancer_type,
             Cancer_type_last_level,
             Specimen_collection_date,
             Panel_name,
             Database_update_date) %>%
    summarise(across(c(Evidence_A, Evidence_B, Evidence_C, Evidence_D, Evidence_E),
                     ~paste(na.omit(.), collapse = ", ")))
  organ_actionability_wide_reachability_data3 <- organ_actionability_wide_reachability_data3 %>%
    group_by(Tumor_Sample_Barcode) %>%
    summarise(across(c(Evidence_A_reachability, Evidence_B_reachability, Evidence_C_reachability, Evidence_D_reachability, Evidence_E_reachability),
                     ~paste(na.omit(.), collapse = ", ")))
  organ_actionability_wide_data3 <- organ_actionability_wide_data3 %>% 
    left_join2(organ_actionability_wide_reachability_data3)
  
  actionability_data[[organ]] <- organ_actionability_wide_data2
  actionability_data2[[organ]] <- organ_actionability_wide_data3
}
actionability_data <- bind_rows(actionability_data) %>% 
  ungroup()
actionability_data2 <- bind_rows(actionability_data2) %>% 
  ungroup()
```

```{r add information}
merged_data <- read_csv("result/table/merged/merged_patient_data2.csv") 
basic_data <- merged_data %>% 
  select(Tumor_Sample_Barcode, Registration_date, 
         Oncotree_CODE, Oncotree_CODE_name, ONCOTREE_Level1, 
         Oncotree_CODE_Final_narrow,
         Oncotree_CODE_Final_wide,
         Patho_diagnosis, Patho_diagnosis_sample)
  
merged_data2 <- merged_data %>% 
  select(-Cancer_type, -Panel_name) %>% 
  left_join2(actionability_data2) %>% 
  mutate(Highest_evidence_level = case_when(
    Evidence_A != "" ~ "Evidence_A",
    Evidence_B != "" ~ "Evidence_B",
    Evidence_C != "" ~ "Evidence_C",
    Evidence_D != "" ~ "Evidence_D",
    Evidence_E != "" ~ "Evidence_E",
    TRUE ~ "No_evidence"
  ))
merged_data2 %>% 
  count(Highest_evidence_level)

actionability_data <- actionability_data %>%
  left_join2(basic_data)
actionability_data2 <- actionability_data2 %>%
  left_join2(basic_data)
```

```{r count}
merged_data2 <- merged_data2 %>%
  mutate(Highest_evidence_level2 = case_when(str_detect(Evidence_A_reachability, "pmda_approved_drug") ~ "Evidence_A(approved)",
                                             TRUE ~ Highest_evidence_level)) %>% 
  mutate(Highest_evidence_level2 = case_when(Highest_evidence_level2 == "Evidence_A" ~ "Evidence_A(unapproved)", 
                                             TRUE ~ Highest_evidence_level2))

merged_data2 %>% 
  count(Highest_evidence_level, Highest_evidence_level2)
```

```{r save}
merged_data2 %>% 
  write_excel_csv("result/table/merged/merged_patient_data3.csv", na = "")
actionability_data %>% 
  write_excel_csv("result/table/merged/merged_actionability_data_by_alteration.csv", na = "")
actionability_data2 %>% 
  write_excel_csv("result/table/merged/merged_actionability_data_by_individual.csv", na = "")
```

```{r sessioninfo}
sessionInfo()
```
