==> 1_preparation/06_postEP_treatment_strategy_data.Rmd <==
---
title: "PostEP_treatment_strategy_data"
author: "Yuki Saito"
date: "2024-03-07"
output: html_document
---

- R markdown script to annotate post-CGP treatment strategies.

```{r setup, include=FALSE, message=FALSE, warning=FALSE}
knitr::opts_chunk$set(echo = TRUE)
knitr::opts_knit$set(root.dir = rprojroot::find_rstudio_root_file())
knitr::opts_chunk$set(root.dir = rprojroot::find_rstudio_root_file())
```

```{r library, message=FALSE, warning=FALSE}
library(tidyverse)
library(dplyr)
library(readxl)
library(lubridate)
```

```{r source}
source("../../R/utils.R")
```

```{r read in merged data, warning=FALSE, message=FALSE}
# read in merged data
merged_data <- read_csv("result/table/merged/merged_patient_data3.csv", na = "")
merged_data

basic_data <- merged_data %>% 
  select(Tumor_Sample_Barcode, Registration_date, 
         Oncotree_CODE, Oncotree_CODE_name, ONCOTREE_Level1, 
         Oncotree_CODE_Final_narrow,
         Oncotree_CODE_Final_wide,
         Patho_diagnosis, Patho_diagnosis_sample)
```

```{r preparation}
# list up organ names
organ_names <- list.files("../../../sh_c-cat_analysis/data/ccat/ver20240621/F1CDx/",
                          full.names = FALSE)
organ_names[1:3]

# colnames reference
case_colnames_df <- read_xlsx("../../ref/colnames_ver202406.xlsx", "case") %>% 
  filter(Post_treatment_columns == "TRUE")
head(case_colnames_df)

case_cols <- case_colnames_df
case_cols["coltypes"][case_cols["coltypes"] == "character"] <- "c"
case_cols["coltypes"][case_cols["coltypes"] == "numeric"] <- "n"
case_cols["coltypes"][case_cols["coltypes"] == "logical"] <- "l"
case_cols["coltypes"][case_cols["coltypes"] == "date"] <- "D"
case_cols_collapsed <- paste(case_cols$coltypes, collapse = "")
```

```{r analysis, warning=FALSE, message=FALSE}
post_treatment_data <- list()
for (organ in organ_names){
  # read in post-treatment data
  organ_post_treatment_data <- read_csv(paste0("result/table/prep/case/", organ, "/post_treatment_data.csv"),
                                        col_types = case_cols_collapsed)
  post_treatment_data[[organ]] <- organ_post_treatment_data
}
post_treatment_data <- bind_rows(post_treatment_data)
```

```{r fill in postEP treatment policy name}
post_treatment_data %>% 
  count(PostEP_treatment_policy, PostEP_treatment_policy_name)

post_treatment_data <- post_treatment_data %>% 
  mutate(PostEP_treatment_policy_name = case_when(PostEP_treatment_policy == 1 ~ "Clinical_trial",
                                                  PostEP_treatment_policy == 4 ~ "Other off label",
                                                  PostEP_treatment_policy == 6 ~ "BSC",
                                                  TRUE ~ PostEP_treatment_policy_name)) 
post_treatment_data %>%
  count(PostEP_treatment_policy, PostEP_treatment_policy_name)
```

```{r estimate postEP treatment policy}
post_treatment_data %>% 
  count(PostEP_treatment_policy, PostEP_treatment_policy_name)

post_treatment_data %>% 
  count(EP_administration_name, PostEP_regimen_nonuse_cause_name, PostEP_treatment_policy, PostEP_treatment_policy_name)

post_treatment_data %>% 
  filter(PostEP_regimen_drug_JP == "治験") %>% 
  count(EP_administration_name, PostEP_regimen_nonuse_cause_name, PostEP_treatment_policy, PostEP_treatment_policy_name)
post_treatment_data %>% 
  filter(PostEP_regimen_drug_JP == "治験") %>% 
  count(PostEP_treatment_policy, PostEP_treatment_policy_name)

post_treatment_data <- post_treatment_data %>% 
  mutate(PostEP_treatment_policy_name_estimated = PostEP_treatment_policy_name) %>% 
  mutate(PostEP_treatment_policy_name_estimated = 
           case_when(PostEP_regimen_drug_JP == "治験" ~ "Clinical_trial",
                     EP_administration_name == "Administered" & PostEP_treatment_policy_name == "Other" ~ "Chemotherapy_given_detail_unknown",
                     PostEP_regimen_nonuse_cause_name == "Chemotherapy other than the indicated drug was given" & is.na(PostEP_treatment_policy_name) ~ "Chemotherapy_given_detail_unknown",
                     PostEP_regimen_nonuse_cause_name == "Chemotherapy other than the indicated drug was given" & PostEP_treatment_policy_name == "Other" ~ "Chemotherapy_given_detail_unknown",
                     PostEP_regimen_nonuse_cause_name == "Chemotherapy was not possible due to poor general condition of the patient" ~ "BSC",
                     PostEP_regimen_nonuse_cause_name == "Chemotherapy was not possible due to the financial circumstances of the patient" ~ "BSC",
                     PostEP_regimen_nonuse_cause_name == "Patient did not wish for chemotherapy" ~ "BSC",
                     PostEP_regimen_nonuse_cause_name == "Death" & is.na(PostEP_treatment_policy_name) ~ "BSC",
                     !is.na(PostEP_regimen_regimen) & is.na(PostEP_treatment_policy_name) ~ "Chemotherapy_given_detail_unknown",
                     !is.na(PostEP_regimen_regimen) & PostEP_treatment_policy_name == "Other" ~ "Chemotherapy_given_detail_unknown",
                     !is.na(`PostEP_regimen_Treatment line_name`) & is.na(PostEP_treatment_policy_name) ~ "Chemotherapy_given_detail_unknown",
                     !is.na(`PostEP_regimen_Treatment line_name`) & PostEP_treatment_policy_name == "Other" ~ "Chemotherapy_given_detail_unknown",
                     !is.na(`PostEP_regimen_drug_input`) & is.na(PostEP_treatment_policy_name) ~ "Chemotherapy_given_detail_unknown",
                     !is.na(`PostEP_regimen_drug_input`) & PostEP_treatment_policy_name == "Other" ~ "Chemotherapy_given_detail_unknown",
                     TRUE ~ PostEP_treatment_policy_name_estimated)) 

post_treatment_data %>%
  count(EP_administration_name, PostEP_regimen_nonuse_cause_name, PostEP_treatment_policy_name, PostEP_treatment_policy_name_estimated) 
post_treatment_data %>% 
  count(PostEP_treatment_policy_name_estimated, PostEP_treatment_policy_name)

post_treatment_data %>% 
  count(PostEP_treatment_policy_name)

post_treatment_data <- post_treatment_data %>% 
  mutate(PostEP_treatment_policy_name = if_else(PostEP_treatment_policy_name %in% c("Corporate clinical trial", "Physician-led clinical trial"), "Clinical_trial", PostEP_treatment_policy_name)) %>%
  mutate(PostEP_treatment_policy_name = if_else(PostEP_treatment_policy_name %in% c("Insurance medical treatment", "Within adaptation"), "Insurance_medical_treatment", PostEP_treatment_policy_name)) %>%
  mutate(PostEP_treatment_policy_name_estimated = if_else(PostEP_treatment_policy_name_estimated %in% c("Corporate clinical trial", "Physician-led clinical trial"), "Clinical_trial", PostEP_treatment_policy_name_estimated)) %>%
  mutate(PostEP_treatment_policy_name_estimated = if_else(PostEP_treatment_policy_name_estimated %in% c("Insurance medical treatment", "Within adaptation"), "Insurance_medical_treatment", PostEP_treatment_policy_name_estimated)) %>%
  mutate(PostEP_treatment_policy_name = if_else(PostEP_treatment_policy_name %in% c("Advanced medical", "Patient offer medical treatment", "Other off label"), "Other_off_label", PostEP_treatment_policy_name)) %>%
  mutate(PostEP_treatment_policy_name_estimated = if_else(PostEP_treatment_policy_name_estimated %in% c("Advanced medical", "Patient offer medical treatment"), "Other_off_label", PostEP_treatment_policy_name_estimated)) %>%
  mutate(PostEP_treatment_policy_name = str_replace_all(PostEP_treatment_policy_name, " ", "_")) %>%
  mutate(PostEP_treatment_policy_name_estimated = str_replace_all(PostEP_treatment_policy_name_estimated, " ", "_")) 

post_treatment_data %>%
  count(PostEP_treatment_policy_name_estimated, PostEP_treatment_policy_name)
```

```{r check}
# check
post_treatment_data %>% 
  filter(PostEP_treatment_policy_name_estimated == "Clinical_trial" & PostEP_treatment_policy_name == "Other_off_label")
post_treatment_data %>% 
  filter(PostEP_treatment_policy_name_estimated == "Clinical_trial" & PostEP_treatment_policy_name == "Insurance medical treatment")
```

```{r save}
post_treatment_data %>% 
  left_join2(basic_data) %>%
  write_excel_csv("result/table/merged/postEP_treatment_data.csv", na = "")
```

```{r summarize into individual level}
post_treatment_data %>% 
  count(PostEP_treatment_policy_name)
post_treatment_data %>% 
  count(PostEP_treatment_policy_name_estimated)

post_treatment_policy_by_individual_data <- post_treatment_data %>%
  select(Tumor_Sample_Barcode, PostEP_treatment_policy_name) %>% 
  group_by(Tumor_Sample_Barcode) %>% 
  summarise(PostEP_treatment_policy_name = case_when(
    any(PostEP_treatment_policy_name == "Clinical_trial", na.rm = TRUE) ~ "Clinical_trial",
    any(PostEP_treatment_policy_name == "Other_off_label", na.rm = TRUE) ~ "Other_off_label",
    any(PostEP_treatment_policy_name == "Insurance_medical_treatment", na.rm = TRUE) ~ "Insurance_medical_treatment",
    any(PostEP_treatment_policy_name == "BSC", na.rm = TRUE) ~ "BSC",
    any(PostEP_treatment_policy_name == "Other", na.rm = TRUE) ~ "Other",
    TRUE ~ NA_character_
  )) 
post_treatment_policy_by_individual_data %>% 
  count(PostEP_treatment_policy_name)

post_treatment_policy_estimated_by_individual_data <- post_treatment_data %>%
  select(Tumor_Sample_Barcode, PostEP_treatment_policy_name_estimated) %>% 
  group_by(Tumor_Sample_Barcode) %>% 
  summarise(PostEP_treatment_policy_name_estimated = case_when(
    any(PostEP_treatment_policy_name_estimated == "Clinical_trial", na.rm = TRUE) ~ "Clinical_trial",
    any(PostEP_treatment_policy_name_estimated == "Other_off_label", na.rm = TRUE) ~ "Other_off_label",
    any(PostEP_treatment_policy_name_estimated == "Insurance_medical_treatment", na.rm = TRUE) ~ "Insurance_medical_treatment",
    any(PostEP_treatment_policy_name_estimated == "Chemotherapy_given_detail_unknown", na.rm = TRUE) ~ "Chemotherapy_given_detail_unknown",
    any(PostEP_treatment_policy_name_estimated == "BSC", na.rm = TRUE) ~ "BSC",
    TRUE ~ NA_character_
  )) 
post_treatment_policy_estimated_by_individual_data %>% 
  count(PostEP_treatment_policy_name_estimated)
```

```{r analysis}
post_treatment_policy_estimated_by_individual_data <- post_treatment_policy_estimated_by_individual_data %>% 
  mutate(PostEP_treatment_policy_name_estimated_rough = if_else(PostEP_treatment_policy_name_estimated == "BSC", "BSC", "Chemotherapy")) %>%
  mutate(PostEP_treatment_policy_name_estimated_detail = case_when(PostEP_treatment_policy_name_estimated == "BSC" ~ "BSC", 
                                                                   PostEP_treatment_policy_name_estimated == "Clinical_trial" ~ "Clinical_trial",
                                                                   PostEP_treatment_policy_name_estimated == "Other_off_label" ~ "Other_off_label",
                                                                   is.na(PostEP_treatment_policy_name_estimated_rough) ~ NA_character_,
                                                                   TRUE ~ "Chemotherapy")) %>%
  mutate(PostEP_treatment_policy_name_estimated_detail = fct_relevel(PostEP_treatment_policy_name_estimated_detail, "Chemotherapy", "Clinical_trial", "Other_off_label", "BSC"))

post_treatment_policy_estimated_by_individual_data %>%
  count(PostEP_treatment_policy_name_estimated, PostEP_treatment_policy_name_estimated_rough, PostEP_treatment_policy_name_estimated_detail)
```

```{r save2}
post_treatment_policy_by_individual_data %>% 
  write_excel_csv("result/table/merged/postEP_treatment_policy_by_individual_data.csv", na = "")
post_treatment_policy_estimated_by_individual_data %>%
  write_excel_csv("result/table/merged/postEP_treatment_policy_estimated_by_individual_data.csv", na = "")

merged_data <- merged_data %>% 
  left_join2(post_treatment_policy_by_individual_data, by = "Tumor_Sample_Barcode") %>%
  left_join2(post_treatment_policy_estimated_by_individual_data, by = "Tumor_Sample_Barcode")
merged_data %>% 
  write_excel_csv("result/table/merged/merged_patient_data4.csv", na = "")
```

```{r sessioninfo}
sessionInfo()
```
