==> 1_preparation/08_MSI_and_TMB.Rmd <==
---
title: "MSI_and_TMB"
author: "Yuki Saito"
date: "2024-10-04"
output: html_document
---

- R markdown script to annotate MSI status and TMB levels.

```{r setup, include=FALSE, message=FALSE, warning=FALSE}
knitr::opts_chunk$set(echo = TRUE)
knitr::opts_knit$set(root.dir = rprojroot::find_rstudio_root_file())
knitr::opts_chunk$set(root.dir = rprojroot::find_rstudio_root_file())
```

```{r library, message=FALSE, warning=FALSE}
library(tidyverse)
library(dplyr)
library(readxl)
library(lubridate)
```

```{r source}
source("../../R/utils.R")
```

```{r read in merged data, warning=FALSE, message=FALSE}
# read in merged data
merged_data <- read_csv("result/table/merged/merged_patient_data5.csv", na = "")
merged_data
```

```{r preparation}
# list up organ names
organ_names <- list.files("../../../sh_c-cat_analysis/data/ccat/ver20240621/F1CDx/",
                          full.names = FALSE)
organ_names[1:3]

# colnames reference
report_colnames_df <- read_xlsx("../../ref/colnames_ver202406.xlsx", "report") %>% 
  filter(colnames_ys %in% c("Tumor_Sample_Barcode", "Cancer_type", "Cancer_type_last_level", "Specimen_collection_date", "Panel_name",
                            "Alteration_type", "Somatic_or_Germline", "DNA_or_RNA", "Marker", "Mutation_context", "Marker_detail",
                            "TMB", "TMB_total", "MSI", "LOH", "Database_update_date"))
head(report_colnames_df)

report_cols <- report_colnames_df
report_cols["coltypes"][report_cols["coltypes"] == "character"] <- "c"
report_cols["coltypes"][report_cols["coltypes"] == "numeric"] <- "n"
report_cols["coltypes"][report_cols["coltypes"] == "logical"] <- "l"
report_cols["coltypes"][report_cols["coltypes"] == "date"] <- "D"
report_cols_collapsed <- paste(report_cols$coltypes, collapse = "")
```

```{r biomarker data, warning=FALSE, message=FALSE}
biomarker_data <- list()
for (organ in organ_names){
  # read in actionability files
  organ_biomarker_data <- read_csv(paste0("result/table/prep/report/", organ, "/other_biomarker_data.csv"),
                                       col_types = report_cols_collapsed)
  
  biomarker_data[[organ]] <- organ_biomarker_data
}
biomarker_data <- bind_rows(biomarker_data)
```

```{r MSI}
msi_biomarker_data <- biomarker_data %>% 
  filter(Marker == "MSI") %>% 
  select(Tumor_Sample_Barcode, Mutation_context) %>% 
  rename(MSI_F1 = Mutation_context)
msi_biomarker_data
```

```{r TMB}
biomarker_data %>% 
  filter(Marker == "TMB") %>%
  count(TMB, sort = TRUE)

# check
biomarker_data %>% 
  filter(Marker == "TMB") %>% 
  filter(str_detect(TMB, "\\*Cannot Be Determined")) %>% 
  count(TMB)
biomarker_data %>% 
  filter(Marker == "TMB") %>%
  filter(TMB %in% c("10mutations-per-megabase(high)*Cannot Be Determined", 
                    "11mutations-per-megabase(high)*Cannot Be Determined",
                    "12mutations-per-megabase(high)*Cannot Be Determined",
                    "18mutations-per-megabase(high)*Cannot Be Determined",
                    "19mutations-per-megabase(high)*Cannot Be Determined"))


TMB_biomarker_data <- biomarker_data %>% 
  filter(Marker == "TMB") %>%
  mutate(TMB_score = str_replace_all(str_replace_all(str_replace_all(str_replace_all(TMB, "mutations-per-megabase", "Muts/Mb"), "Muts/Mb\\(\\)", ""), "Muts/Mb\\(high\\)", ""), "Muts/Mb\\(low\\)", "")) %>% 
  mutate(TMB_score = str_replace_all(TMB_score, "Muts/Mb", "")) %>% 
  mutate(TMB_score = if_else(str_detect(TMB_score, "Cannot Be Determined"), NA_character_, TMB_score))
  #mutate(TMB_score = str_replace_all(TMB_score, "\\*Cannot Be Determined", ""))

TMB_biomarker_data <- TMB_biomarker_data %>% 
  select(Tumor_Sample_Barcode, Mutation_context, TMB_score, TMB) %>% 
  mutate(TMB_score = as.numeric(TMB_score))
TMB_biomarker_data
```

```{r summarize}
TMB_biomarker_data %>% 
  filter(Mutation_context == "high") %>% 
  summary(TMB_score)

tmp <- TMB_biomarker_data %>% 
  filter(is.na(Mutation_context)) %>%
  filter(TMB_score >= 10)
tmp

merged_data %>% 
  filter(Tumor_Sample_Barcode %in% tmp$Tumor_Sample_Barcode) %>% 
  select(Evidence_A, Evidence_B, Evidence_C, Evidence_D, Evidence_E) %>% 
  filter(str_detect(Evidence_A, "TMB") |
           str_detect(Evidence_B, "TMB") |
           str_detect(Evidence_C, "TMB") |
           str_detect(Evidence_D, "TMB") |
           str_detect(Evidence_E, "TMB"))

TMB_biomarker_data <- TMB_biomarker_data %>% 
  rename(TMB_F1 = Mutation_context) %>% 
  rename(TMB_score_F1 = TMB_score)
```

```{r save}
msi_biomarker_data %>% 
  write_excel_csv("result/table/merged/MSI_F1_data.csv", na = "")
TMB_biomarker_data %>% 
  write_excel_csv("result/table/merged/TMB_F1_data.csv", na = "")

merged_data <- merged_data %>% 
  left_join2(msi_biomarker_data)  %>% 
  mutate(MSI_EvidenceA = case_when(str_detect(Evidence_A_reachability, "MSI_high_Somatic") ~ "MSI_high",
                                   TRUE ~ "")) %>% 
  left_join2(TMB_biomarker_data) %>% 
  mutate(TMB_EvidenceA = case_when(str_detect(Evidence_A_reachability, "TMB_high_Somatic") ~ "TMB_high",
                                   TRUE ~ ""))
merged_data %>% 
  write_excel_csv("result/table/merged/merged_patient_data6.csv", na = "")
```

```{r sessioninfo}
sessionInfo()
```
