==> 1_preparation/09_number_of_alterations.Rmd <==
---
title: "Number of alterations"
author: "Yuki Saito"
date: "2024-03-07"
output: html_document
---

-  R markdown script to annotate the number of genetic alterations by case.

```{r setup, include=FALSE, message=FALSE, warning=FALSE}
knitr::opts_chunk$set(echo = TRUE)
knitr::opts_knit$set(root.dir = rprojroot::find_rstudio_root_file())
knitr::opts_chunk$set(root.dir = rprojroot::find_rstudio_root_file())
```

```{r library, message=FALSE, warning=FALSE}
library(tidyverse)
library(dplyr)
library(readxl)
library(lubridate)
```

```{r source}
source("../../R/utils.R")
```

```{r read in merged data, warning=FALSE, message=FALSE}
# read in merged data
merged_data <- read_csv("result/table/merged/merged_patient_data6.csv", na = "")
merged_data
```

```{r preparation}
# list up organ names
organ_names <- list.files("../../../sh_c-cat_analysis/data/ccat/ver20240621/F1CDx/",
                          full.names = FALSE)
organ_names[1:3]
```

```{r numbers, warning=FALSE, message=FALSE}
# colnames reference
report_colnames_df <- read_xlsx("../../ref/colnames_ver202406.xlsx", "report") %>% 
  filter(Number_columns == "TRUE")
head(report_colnames_df)

report_cols <- report_colnames_df
report_cols["coltypes"][report_cols["coltypes"] == "character"] <- "c"
report_cols["coltypes"][report_cols["coltypes"] == "numeric"] <- "n"
report_cols["coltypes"][report_cols["coltypes"] == "logical"] <- "l"
report_cols["coltypes"][report_cols["coltypes"] == "date"] <- "D"
report_cols_collapsed <- paste(report_cols$coltypes, collapse = "")

number_data <- list()
for (organ in organ_names){
  # read in post-treatment data
  organ_number_data <- read_csv(paste0("result/table/prep/report/", organ, "/number_data.csv"),
                                       col_types = report_cols_collapsed)
  number_data[[organ]] <- organ_number_data
}
number_data <- bind_rows(number_data)
```

```{r numbers2, warning=FALSE, message=FALSE}
number_data2 <- list()
for (organ in organ_names){
  # read in post-treatment data
  SNV_data <- read_csv(paste0("result/table/prep/report/", organ, "/small_scale_variant_data.csv")) %>% 
    count(Tumor_Sample_Barcode) %>% 
    rename(SNV_num = n)
  CNV_data <- read_csv(paste0("result/table/prep/report/", organ, "/CNV_data.csv")) %>% 
    count(Tumor_Sample_Barcode) %>% 
    rename(CNV_num = n)
  SV_data <- read_csv(paste0("result/table/prep/report/", organ, "/rearrangement_data.csv")) %>%
    count(Tumor_Sample_Barcode) %>% 
    rename(Rearrangement_num = n)
  organ_number_data2 <- SNV_data %>% 
    full_join(CNV_data, by = "Tumor_Sample_Barcode") %>%
    full_join(SV_data, by = "Tumor_Sample_Barcode") %>% 
    mutate(across(everything(), ~replace_na(., 0)))
  number_data2[[organ]] <- organ_number_data2
}
number_data2 <- bind_rows(number_data2)
```

```{r merge, message=FALSE, warning=FALSE}
number_data_summary <- number_data %>% 
  select(Tumor_Sample_Barcode, Num_of_somatic_alts) %>%
  left_join2(number_data2) %>% 
  mutate(across(everything(), ~replace_na(., 0)))
print(number_data_summary)
```

```{r save}
merged_data <- merged_data %>% 
  left_join2(number_data_summary, by = "Tumor_Sample_Barcode")
merged_data %>% 
  write_excel_csv("result/table/merged/merged_patient_data7.csv", na = "")
```

```{r sessioninfo}
sessionInfo()
```
