==> 1_preparation/13_double_cancer.Rmd <==
---
title: "Double_cancer"
author: "Yuki Saito"
date: "2025-07-21"
output: html_document
---

- R markdown script to annotate the presence of double cancers for each case.

```{r setup, include=FALSE, message=FALSE, warning=FALSE}
knitr::opts_chunk$set(echo = TRUE)
knitr::opts_knit$set(root.dir = rprojroot::find_rstudio_root_file())
knitr::opts_chunk$set(root.dir = rprojroot::find_rstudio_root_file())
```

```{r library, message=FALSE, warning=FALSE}
library(tidyverse)
library(dplyr)
library(readxl)
library(lubridate)
library(cowplot)
library(stringi)
library(readr)
theme_set(theme_cowplot())
```

```{r source}
source("../../R/utils.R")
source("../../R/figure.R")
source("../../R/automatic_annotation.R")
```

```{r preparation for double cancer}
# list up organ names
organ_names <- list.files("../../../sh_c-cat_analysis/data/ccat/ver20240621/F1CDx/",
                          full.names = FALSE)
organ_names[1:3]

# colnames reference
report_colnames_df <- read_xlsx("../../ref/colnames_ver202406.xlsx", "case") %>% 
  filter(Double_multiple_columns == "TRUE")
head(report_colnames_df)

report_cols <- report_colnames_df
report_cols["coltypes"][report_cols["coltypes"] == "character"] <- "c"
report_cols["coltypes"][report_cols["coltypes"] == "numeric"] <- "n"
report_cols["coltypes"][report_cols["coltypes"] == "logical"] <- "l"
report_cols["coltypes"][report_cols["coltypes"] == "date"] <- "D"
report_cols_collapsed <- paste(report_cols$coltypes, collapse = "")
```

```{r double cancer, warning=FALSE, message=FALSE}
multiple_data <- list()
for (organ in organ_names){
  # read in actionability files
  organ_multiple_data <- read_csv(paste0("result/table/prep/case/", organ, "/double_multiple_data.csv"),
                                       col_types = report_cols_collapsed)
  
  multiple_data[[organ]] <- organ_multiple_data
}
multiple_data <- bind_rows(multiple_data) 

multiple_data %>% 
  count(Double_cancer, Double_cancer_name, Double_cancer_activity, Double_cancer_activity_name)

multiple_data <- multiple_data %>% 
  select(Tumor_Sample_Barcode, Double_cancer_name, Double_cancer_activity_name) %>% 
  unique()
multiple_data

Double_cancer_data <- multiple_data %>% 
  filter(Double_cancer_name %in% c("Yes", "None")) %>% 
  select(Tumor_Sample_Barcode, Double_cancer_name) %>%
  unique()

# check
tmp <- Double_cancer_data %>% 
  count(Tumor_Sample_Barcode) %>% 
  filter(n >= 2)
if (nrow(tmp) > 0){
  stop("something wrong")
}

Double_cancer_activity_data <- multiple_data %>% 
  filter(!is.na(Double_cancer_name)) %>% 
  mutate(Double_cancer_activity_name = paste(Double_cancer_name, Double_cancer_activity_name, sep = "_")) %>% 
  select(Tumor_Sample_Barcode, Double_cancer_activity_name) %>%
  unique()
Double_cancer_activity_data
```

```{r read in merged data, warning=FALSE, message=FALSE}
# read in merged data
merged_data <- read_csv("result/table/merged/merged_patient_data9.csv", na = "")

merged_data2 <- merged_data %>% 
  left_join2(Double_cancer_data)

merged_data2 %>% 
  count(Double_cancer_name)
```

```{r save}
merged_data2 %>% 
  write_excel_csv("result/table/merged/merged_patient_data10.csv", na = "")
```

```{r sessioninfo}
sessionInfo()
```
